Implementa SOLID ao Menu
